+++ platformio/program_esp32/src/ADSConverter.cpp
@@ -14,11 +14,11 @@ bool ADSInit::init() {
  * 3. adc result
  * 4. voltage result
 */
-void SEN_0189::run(uint8_t pin, long delayms, int16_t* adc_ads, float* volt_ads) {
+void SEN_0189::run(uint8_t pin, long delayms, int* adc_ads, float* volt_ads) {
     if ((unsigned long) (millis() - LastDelay) >= (long unsigned int) delayms) {
         LastDelay = millis();
         
-        int16_t _adc_ads = ADSInit::ads.readADC_SingleEnded(pin);
+        int _adc_ads = ADSInit::ads.readADC_SingleEnded(pin);
         float _volt = ADSInit::ads.computeVolts(_adc_ads);
 
         *adc_ads = _adc_ads;
@@ -26,11 +26,11 @@ void SEN_0189::run(uint8_t pin, long delayms, int16_t* adc_ads, float* volt_ads)
     }
 }
 
-void SEN_0257::run(uint8_t pin, long delayms, int16_t* adc_ads, float* volt_ads) {
+void SEN_0257::run(uint8_t pin, long delayms, int* adc_ads, float* volt_ads) {
     if ((unsigned long) (millis() - LastDelay) >= (long unsigned int) delayms) {
         LastDelay = millis();
 
-        int16_t _adc_ads = this->ads.readADC_SingleEnded(pin);
+        int _adc_ads = this->ads.readADC_SingleEnded(pin);
         float _volt = this->ads.computeVolts(_adc_ads);
 
         *adc_ads = _adc_ads;
